requirements.txt
python-dotenv==1.2.1
pydantic==2.12.5
python-multipart==0.0.21
starlette==0.50.0
uvicorn==0.40.0
fastapi==0.128.0
openai==2.15.0
chromadb==1.4.1
pandas==2.3.3
tqdm==4.67.1
langgraph==1.0.7
structlog==25.5.0
pypdf2==3.0.1
python-docx==1.2.0
openpyxl==3.1.5